Update problem.jl
I want to introduce different type of solvers, not necessarily explicit.
Such solvers will typically not depend on f!, but on a suitable decomposition (for instance exponential solvers). Hence the step should depend on the model itself (which should be suitably defined, hence the "basic check"), and not on model.f!